Add some description files for flavours.  Basically, for each flavour, if 'class' is set, then in the description, "@class@ class systems" will be described; if class is not set, then the flavour name will be used.
Ie, if there's no desc.386, then "this kernel is for 386 class systems" will
show up in the description.

There's also @desc@ (see the s390 stuff for an example) that will be appended
to the description.

I haven't actually implemented this yet; please fill in what you need for
your pet architecture, and I'll do it tomorrow.


svn path=/trunk/kernel/source/linux-2.6-2.6.12/; revision=3594