Really fix missing kernel-img.conf makes a debian kernel not installable
There is no /etc/kernel-img.con on a fresh debootstrap sid. Installing a
kernel leads to:

|mikejones:/var/lib# dpkg --configure -a
|Setting up linux-image-2.6.31-1-amd64 (2.6.31-1) ...
|Running depmod.
|Running update-initramfs.
|update-initramfs: Generating /boot/initrd.img-2.6.31-1-amd64
|Error retreiving answer for linux-image-2.6.31-1-amd64/postinst/create-kimage-link-2.6.31-1-amd64:
| linux-image-2.6.31-1-amd64/postinst/create-kimage-link-2.6.31-1-amd64 doesn't exist at
| /var/lib/dpkg/info/linux-image-2.6.31-1-amd64.postinst line 522, <STDIN> line 3.
|dpkg: error processing linux-image-2.6.31-1-amd64 (--configure):
| subprocess installed post-installation script returned error exit status 9

As far as I understand debconf, there must be a template and input() for
${package_name}/postinst/create-kimage-link-$version in order to get a
value saved. Since I did not find it I guess that this can't work.

If the bootloader is palo than the answer is ignored anyway so the
question is not required there.

kernel-img.conf(5) says that do_symlinks defaults to yes.

This could also fix piuparts [0].

[0] http://piuparts.debian.org/sid/fail/linux-image-2.6.31-1-amd64_2.6.31-1.log

[ nuked that template in question and this is indeed the right fix
  instead of reprompting user for useless questions.
  Add a comment and keep current ordering. -maks ]

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>

Really fix making a debian kernel installable without kernel-img.conf.

boah the code below this with reverse links must die pretty soon.

thanks for good analysis and bug report.

svn path=/dists/trunk/linux-2.6/; revision=14589