Deal with ABI changes in 3.1.5
struct clocksource changed but clocksource functions shouldn't be used
by OOT modules; ignore the change.

struct inet_peer changed; although it shouldn't be used by OOT modules
it does affect symbol versions for many of the networking core
functions.  Since the structure is only allocated in one place so we
can safely extend it instead of adding the new member in the middle.

svn path=/dists/sid/linux-2.6/; revision=18372