Add more kluges to the cross-modpost makefile for 3.9
mod_devicetable.h is now converted into a series of offset definitions
in devicetable-offsets.h which are included by file2alias.c.  So we
have to replicate that as well.

Note that there are some architecture combinations for which some
alias types won't be parsed correctly (and probably never were).
The current cross-modpost build rules are really not sustainable.

svn path=/dists/trunk/linux-tools/; revision=20135