[armel,armhf] linux-tools: Enable use of libunwind
perf requires some architecture-specific code in conjunction with
libunwind.  It used to enable use of libunwind automatically anyway,
so we had to override this for architectures where it would fail to
build (by setting NO_LIBUNWIND=1).

The makefile is now fixed so we don't need to this.  Since perf now
has the extra code for arm, this enables use of libunwind there.

svn path=/dists/sid/linux-tools/; revision=21910