[arm64,armhf,powerpcspe,x86] gpio: Explicitly enable GPIOLIB
Some drivers now depend on rather than selecting it, in particular
FB_VIA.  This fixes an FTBFS on i386 as that driver was no longer
built on 686-pae.

The architectures are chosen to cover everything that selects
ARCH_WANT_OPTIONAL_GPIOLIB and for which we have GPIOLIB enabled
in jessie.

svn path=/dists/trunk/linux/; revision=22564