linux-source: Fix compression of the source tarball (Closes: #795199)
When we introduced the find | sort | xargs tar pipeline to make the
tarball reproducible, I was concerned that there could be so many
arguments that xargs would call tar more than once.  I therefore
changed the tar -c option to -r so that it would append rather
than overwriting in this case.

However, the -a option seems to have no effect when -r is used, so
the tarball is no longer compressed.

Instead of using xargs, use tar's -T option and change the -r
back to -c.