Add config parameter to disable linux-headers-all packages
The linux-grsec source package needs a way to explicitly disable these
binary packages which are already built by the linux source package.
We already do that when there are no actual kernels for the target
architecture.  Rename the FOREIGN_KERNEL make variable and combine the
two conditions.

Based on work by Yves-Alexis Perez.