[mipsel/mips/config.loongson-2f] Disable VIDEO_CX23885, VIDEO_IVTV, VIDEO_CX231XX, VIDEO_PVRUSB2 (fixes FTBFS).
This workaround the failed compilation of cx25840-core.c:

| {standard input}: Assembler messages:
| {standard input}:583: Error: branch out of range
| {standard input}:636: Error: branch out of range
| {standard input}:40094: Error: branch out of range

These are caused by the -Wa,-mfix-loongson2f-jump, which replaces the
"jalr t9" instruction by the "or t9, t9, 0x80000000; jalr t9"
instructions, thus causing an increase in code size.

This is probably the best way to workaround the FTBFS, especially given
that the loongson-2e/2f flavour are likely going to be removed for
stretch.