debian/rules.d/Makefile: Override architecture detection when building userland headers
Martin Michlmayr found that userland builds fail if 'uname -m' reports
the native kernel architecture, and that's because we don't explicitly
set ARCH for the headers_install target.  That's fine for most cases
of a 32-bit chroot on a 64-bit kernel because a single source
architecture covers both, but arm/arm64 is an exception.

This doesn't make actual cross-builds of userland tools work.