usbip: Use the locally-installed <linux/usbip.h>
Some years back I patched the #include of <linux/usbip.h> to include
the current, uninstalled version through a relative path because the
system-installed version will normally be outdated.  We're about to
start copying the source to avoid writing autotools crap into the
source directory, which breaks this relative path.

Since we now install the UAPI headers under debian/build at the start
of the tools build, drop the patch and add that directory to the
include path.