debian/rules.real: Avoid error message when usbip's config.h doesn't exist
usbip has its own version number which we combine with the source
package version, which is assigned to VERSION for the install-usbip
target (only).  We find the version number by processing the config.h
file created by autoconf.  The file always exists before the
install-usbip rule is invoked, but the target-specific definition of
VERSION is still evaluated whenever debian/rules.real is used,
resulting in confusing (though harmless) error messages about a
missing file.

We could change VERSION to be a recursively-expanded variable, but
then it would still be evaulated multiple times.  Instead, move the
definition of VERSION into the target's commands.