Revert "arm64/mm: Limit TASK_SIZE_64 ..." and add breaks on incompatible mozjs
mozjs assumed VAs would never be wider than 47 bits.  Add Breaks to
force upgrades of those packages.  For mozjs 1.8.5 the fix requires an
ABI change so this is unversioned; for mozjs 24 there was no ABI
change so this is versioned.

(luajit has the same problem but only the experimental version
supports arm64.  I assume this will be fixed before it goes into
unstable, so we don't need it in Breaks.)