time: Disable TIMER_STATS (CVE-2017-5967)
The upstream "fix" for this is to remove the feature, as it is
redundant with tracing.  I'd be quite happy to do that, but it
introduces several conflicts with the PREEMPT_RT patch series.

Unless and until those are resolved in 4.9-stable and 4.9-rt, disable
it in our kconfig and add a dependency on BROKEN to ensure it's
disabled in custom kernels too.

(cherry picked from commit 49c2b929374d04186ea2ac402f045094516630e9)