[sparc64] Update "Revert "sparc: move exports to definitions"" again
This should really fix the FTBFS - at least, the build got as far as
building linux-image packages.

genksyms doesn't recognise __int128 as a type name, so fails to parse
the prototype for __multi3().  I could fix genksyms but would have to
regenerate the parser tables which would be a horrible patch to
maintan.  So use a struct type instead for now.  gcc doesn't seem to
care about this because it isn't a normal C function.

Also update the patch properly for 4.12 - I removed exports for some
symbols that were not really removed but renamed.