linux-image: Recommend apparmor (Closes: #880441)
The apparmor package is needed for loading profiles.  In theory,
enabling AppArmor without any profiles loaded should do nothing, so
this is not really a dependency of the kernel.  However, if a systemd
unit specifies a AppArmor profile and the kernel has AppArmor enabled
then failure to load the profile is fatal.

As the linux-image packages select AppArmor as the default LSM, they
should probably also be responsible for getting the necessary userland
support installed.  But since the default can be overridden, use
Recommends rather than Depends.