Bump ABI to 2
Commit 1df9e416e647 "Kbuild: always define endianess in kconfig.h"
fixed several cases where some types (and inline functions) might not
be correctly defined according to the host byte order.  It might be
possible to avoid an ABI bump, but it would require a lot of work.
Also, the problem may affect OOT modules that therefore should be
rebuilt with the fix.